+++ service-2/app.py
@@ -3,10 +3,11 @@ import random
 
 app = Flask(__name__)
 
-@app.route('/get_coords', methods=['GET'])
+@app.route('/get_coords', methods=['GET', 'POST'])
 def coords():
     x = str(random.randrange(-85, 85))
     y = str(random.randrange(-180, 180))
     return x + ", " + y
 
-if __name__ == "__main__": app.run(host="0.0.0.0", port=5000, debug=True)
+if __name__ == "__main__": 
+    app.run(host="0.0.0.0", port=5000, debug=True)
